+++ segment.py
@@ -24,7 +24,7 @@ data_feed = Data()
 
 batch_data = tf.placeholder(name='BatchData', dtype=tf.float64,
                             shape=[None, data_feed.image_height, data_feed.image_width, data_feed.image_depth])
-true_segmentation = tf.placeholder(name='TrueSegmentation', dtype=tf.float64,
+true_segmentation = tf.placeholder(name='TrueSegmentation', dtype=tf.int32,
                                    shape=[None, data_feed.image_height, data_feed.image_width, 1])
 
 output, optimize, loss = build_model(image_batch=batch_data, true_segmentation=true_segmentation)
